+++ src/posthog-core.ts
@@ -1096,7 +1096,7 @@ export class PostHog {
      *     // Display the properties
      *     console.log(posthog.persistence.properties())
      *
-     * @param {Object} properties An associative array of properties to store about the user
+     * @param {Object} properties properties to store about the user
      * @param {Number} [days] How many days since the user's last visit to store the super properties
      */
     register(properties: Properties, days?: number): void {
@@ -1604,6 +1604,13 @@ export class PostHog {
                 ''
             )
         }
+
+        this.register(
+            {
+                $last_posthog_reset: new Date().toISOString(),
+            },
+            1
+        )
     }
 
     /**
@@ -1948,13 +1955,16 @@ export class PostHog {
     captureException(error: Error, additionalProperties?: Properties): void {
         const syntheticException = new Error('PostHog syntheticException')
         const properties: Properties = isFunction(assignableWindow.__PosthogExtensions__?.parseErrorAsProperties)
-            ? assignableWindow.__PosthogExtensions__.parseErrorAsProperties(
-                  [error.message, undefined, undefined, undefined, error],
-                  // create synthetic error to get stack in cases where user input does not contain one
-                  // creating the exceptionas soon into our code as possible means we should only have to
-                  // remove a single frame (this 'captureException' method) from the resultant stack
-                  { syntheticException }
-              )
+            ? {
+                  ...assignableWindow.__PosthogExtensions__.parseErrorAsProperties(
+                      [error.message, undefined, undefined, undefined, error],
+                      // create synthetic error to get stack in cases where user input does not contain one
+                      // creating the exceptions soon into our code as possible means we should only have to
+                      // remove a single frame (this 'captureException' method) from the resultant stack
+                      { syntheticException }
+                  ),
+                  ...additionalProperties,
+              }
             : {
                   $exception_level: 'error',
                   $exception_list: [